fix test code and readme. ordered build rules.
