+++ CoreData/Tests/Planet+CoreDataProperties.swift
@@ -1,8 +1,8 @@
 //
 //  Planet+CoreDataProperties.swift
-//  BarrelCoreData
+//  Barrel
 //
-//  Created by Nobuo Saito on 2015/11/10.
+//  Created by Nobuo Saito on 2015/11/17.
 //  Copyright © 2015年 tarunon. All rights reserved.
 //
 //  Choose "Create NSManagedObject Subclass…" from the Core Data editor menu
@@ -14,10 +14,8 @@ import CoreData
 
 extension Planet {
 
-    @NSManaged var name: String
-    @NSManaged var diameter: NSNumber
     @NSManaged var semiMajorAxis: NSNumber
-    @NSManaged var parent: Star?
     @NSManaged var children: Set<Satellite>
+    @NSManaged var parent: Star?
 
 }
