+++ CoreData/Tests/Planet.swift
@@ -1,8 +1,8 @@
 //
 //  Planet.swift
-//  BarrelCoreData
+//  Barrel
 //
-//  Created by Nobuo Saito on 2015/11/10.
+//  Created by Nobuo Saito on 2015/11/17.
 //  Copyright © 2015年 tarunon. All rights reserved.
 //
 
@@ -10,7 +10,7 @@ import Foundation
 import CoreData
 
 
-class Planet: NSManagedObject {
+class Planet: StarBase {
 
 // Insert code here to add functionality to your managed object subclass
