+++ CoreData/Tests/Satellite+BarrelAttribute.swift
@@ -10,9 +10,7 @@ import Foundation
 import Barrel
 import Barrel_CoreData
 
-extension AttributeType where FieldType == Satellite {
-    var diameter: Attribute<NSNumber> { return storedAttribute(parent: self) }
-    var name: Attribute<String> { return storedAttribute(parent: self) }
+extension AttributeType where FieldType: Satellite {
     var semiMajorAxis: Attribute<NSNumber> { return storedAttribute(parent: self) }
     var parent: OptionalAttribute<Planet> { return storedAttribute(parent: self) }
 }
