+++ CoreData/Tests/Satellite.swift
@@ -1,8 +1,8 @@
 //
 //  Satellite.swift
-//  BarrelCoreData
+//  Barrel
 //
-//  Created by Nobuo Saito on 2015/11/10.
+//  Created by Nobuo Saito on 2015/11/17.
 //  Copyright © 2015年 tarunon. All rights reserved.
 //
 
@@ -10,7 +10,7 @@ import Foundation
 import CoreData
 
 
-class Satellite: NSManagedObject {
+class Satellite: StarBase {
 
 // Insert code here to add functionality to your managed object subclass
