+++ Realm/Tests/Satellite.swift
@@ -11,17 +11,13 @@ import RealmSwift
 import Barrel
 @testable import Barrel_Realm
 
-class Satellite: Object {
-    dynamic var name: String = ""
-    dynamic var diameter: Double = 0.0
+class Satellite: StarBase {
     dynamic var semiMajorAxis: Double = 0.0
     
     dynamic var parent: Planet?
 }
 
-extension AttributeType where FieldType == Satellite {
-    var name: Attribute<String> { return storedAttribute(parent: self) }
-    var diameter: Attribute<Double> { return storedAttribute(parent: self) }
+extension AttributeType where FieldType: Satellite {
     var semiMajorAxis: Attribute<Double> { return storedAttribute(parent: self) }
     var parent: OptionalAttribute<Planet> { return storedAttribute(parent: self) }
 }
